Create activity_chart_selector
This UI script add a functionality to any activity chart (displaying test statuses). It allows to hide any status from the chart. That is useful because some curves can hide others when they overlap or when values are too high. User must click on a status (on the right side). Clicking again unhide it.